requirements.txt
gradio
requests
openai-whisper
ffmpeg-python